Merged PR 6973517: Prettier - avoid EOL encoding
Current State:
git checks out files with encoding CRLF, and checks back in as LF. This is the default behavior for git in windows.
Prettier defaults to LF. Hence, when we create a branch and execute npm run start --> it automatically formats the code, and converts all the CRLF --> LF which causes noise in commit. Lots of files are unnecessarily added to the commit with "whitespace" being the only difference.

Solution:
set Prettier options to keep existing End Of Line settings as is and not touch it 😊

Impact: this will keep our commits clean, and allow us to trace back through git blame properly rather than introducing noise.

CRLF = Carriage Return Line Feed
LF = Line Feed

Related work items: #38204537