Merged PR 9039818: Correcting the formula
Correcting the formula.

The conversion of date to long is an issue which is Year*10000+Month*100+Date+hour.

If startdate is 05/05/2023 4th hour -> 20230509
enddate is 05/19/2023 4th hour -> 20230523

Now, because of this error some conversion, streams outside the window range were also added to the stream set. Like 05/04/2023 5th hour -> 20230509 is >= startdateid so streams not required will also get considered.

Changing the formula to Year*1000000+Month*10000+Date*100+hour will solve the issue.

Attaching the stats file->
OriginalStats is the stats with the older and error formula calculation -> set of streams getting considered.

OriginalStats11 -> new and correct formula calculation stream set consideration.

Related work items: #44792811