Merged PR 9193387: Remove Generic Failures & Devices Flattened Generic generation
Remove Generic Failures & Devices Flattened Generic generation. Failures Generic was already removed, but now we are removing the device use as well.

Current WF is failing due to Failures Generic being missing. But the outout is not being used in cosmos or kusto, so removing the input for failures flattened.

Related work items: #45220714